requirements.txt
annotated-types==0.7.0
anyio==4.4.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.1
distro==1.9.0
fastapi==0.115.8
firebase-admin==6.2.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.8
jiter==0.5.0
msgpack==1.1.0
openai==1.42.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.40.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==14.0
